+++ tests/test_app/TestApp/Controller/ViewPostsController.php
@@ -1,0 +1,52 @@
+<?php
+declare(strict_types=1);
+/**
+ * CakePHP(tm) : Rapid Development Framework (https://cakephp.org)
+ * Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
+ *
+ * Licensed under The MIT License
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright     Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
+ * @link          https://cakephp.org CakePHP(tm) Project
+ * @since         3.0.0
+ * @license       https://opensource.org/licenses/mit-license.php MIT License
+ */
+namespace TestApp\Controller;
+
+use Cake\Controller\Controller;
+
+class ViewPostsController extends Controller
+{
+    /**
+     * name property
+     *
+     * @var string
+     */
+    public $name = 'Posts';
+
+    /**
+     * index method
+     *
+     * @return void
+     */
+    public function index()
+    {
+        $this->set([
+            'testData' => 'Some test data',
+            'test2' => 'more data',
+            'test3' => 'even more data',
+        ]);
+    }
+
+    /**
+     * nocache_tags_with_element method
+     *
+     * @return void
+     */
+    public function nocache_multiple_element()
+    {
+        $this->set('foo', 'this is foo var');
+        $this->set('bar', 'this is bar var');
+    }
+}
