+++ tests/test_app/TestApp/Utility/TextMock.php
@@ -1,0 +1,7 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Utility;
+
+class TextMock
+{
+}
