+++ tests/test_app/TestApp/View/Helper/HtmlAliasHelper.php
@@ -1,0 +1,19 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\View\Helper;
+
+use Cake\View\Helper;
+
+/**
+ * Extended HtmlHelper
+ */
+class HtmlAliasHelper extends Helper
+{
+    /**
+     * @param string $viewFile
+     * @return void
+     */
+    public function afterRender($viewFile)
+    {
+    }
+}
