+++ tests/test_app/TestApp/View/Helper/TextHelperTestObject.php
@@ -1,0 +1,19 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\View\Helper;
+
+use Cake\View\Helper\TextHelper;
+use TestApp\Utility\TextMock;
+
+class TextHelperTestObject extends TextHelper
+{
+    public function attach(TextMock $string)
+    {
+        $this->_engine = $string;
+    }
+
+    public function engine()
+    {
+        return $this->_engine;
+    }
+}
