+++ tests/test_app/TestApp/View/Object/TestObjectWithToString.php
@@ -1,0 +1,21 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\View\Object;
+
+/**
+ * TestObjectWithToString
+ *
+ * An object with the magic method __toString() for testing with view blocks.
+ */
+class TestObjectWithToString
+{
+    /**
+     * Return string value.
+     *
+     * @return string
+     */
+    public function __toString()
+    {
+        return "I'm ObjectWithToString";
+    }
+}
