+++ tests/test_app/TestApp/View/Object/TestObjectWithoutToString.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\View\Object;
+
+/**
+ * TestObjectWithoutToString
+ *
+ * An object without the magic method __toString() for testing with view blocks.
+ */
+class TestObjectWithoutToString
+{
+}
