+++ tests/test_app/TestApp/View/TestStringTemplate.php
@@ -1,0 +1,17 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\View;
+
+use Cake\Core\InstanceConfigTrait;
+use Cake\View\StringTemplateTrait;
+
+class TestStringTemplate
+{
+    use InstanceConfigTrait;
+    use StringTemplateTrait;
+
+    /**
+     * @var array
+     */
+    protected $_defaultConfig = [];
+}
